Fix readonly file mapping on windows
`CreateFileMapping` tries to extend the file to the size of mapping
which leads to failure if database was opened in readonly and calculated
mmap size is bigger than the file size.
Providing 0 to `MapViewOfFile` will create a view which has size of
mapping, i.e. file-size in read-only mode and full size if file was
truncated.

Also, swap `sizehi` and `sizelo` names to reflect windows API docs.
This was changed in 1c97a490d for seemingly no reason.

Signed-off-by: Evgenii Stratonikov <evgeniy@nspcc.ru>